sync recent miner/ccc changes (#1062)
* refactor: eliminate double re-execution in AsyncChecker (#1036)

* fix: make reorg mode explicit (#1049)

* fix: avoid committing empty blocks after the deadline (#1051)

* fix: initialize pending block with an empty block (#1052)

* fix: reuse timestamp for blocks failing CCC (#1031)

we assume that same height wont trigger multiple reorgs to be able
to put an upper bound on the reorg depth. We rely on the fact that
AsyncChecker executes transactions one-by-one and tells worker the
safe set of transactions to include in the replacement block that
wont trigger another error on the same block. If worker changes the
timestamp and that causes significant changes to the execution flow
of included transactions; we might have a height where multiple
reorgs happen.

---------

Co-authored-by: Ömer Faruk Irmak <omerfirmak@gmail.com>